Move create_app to startup_utils
Stop importing create_app in
datacube_ows to avoid loading
in the entire application
when starting datacube-ows-update.

Refs #1549
