+++ datacube_ows/__init__.py
@@ -11,6 +11,4 @@ except ImportError:
     # Will only be used when running datacube-ows direct from source code (not properly installed)
     __version__ = "1.9.8"
 
-from .startup_utils import create_app
-
-__all__ = ["__version__", "create_app"]
+__all__ = ["__version__"]
